requirements.txt
aniso8601==7.0.0
asgiref==3.2.10
autopep8==1.5.3
dj-database-url==0.5.0
Django==3.0.8
django-ckeditor==5.9.0
django-cors-headers==3.4.0
django-graphql-jwt==0.3.1
django-heroku==0.3.1
django-js-asset==1.2.2
graphene==2.1.8
graphene-django==2.12.1
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.0.4
promise==2.3
psycopg2==2.8.5
pycodestyle==2.6.0
PyJWT==1.7.1
pytz==2020.1
Rx==1.6.1
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
Unidecode==1.1.1
whitenoise==5.1.0
